changed UserContext on pubusercomp to PubUserContext
