+++ src/components/context/UserContext.jsx
@@ -28,32 +28,8 @@ const UserProvider = ({ children }) => {
     }
   };
 
-// ===========================================================
-// GET PUBLIC USER FROM ID
-
-  const getPublicUserFromId = async (userId) => {
-    const options2 = {
-      method: "GET",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      credentials:"include",
-    };
-    try {
-      const response = await fetch(
-        `${import.meta.env.VITE_API_URL}/user/public-user/${userId}`,
-        options2
-      );
-      const fetchData = await response.json();
-      setUserData(fetchData);
-    } catch (error) {
-      console.log("Error fetching: " + error);
-    }
-  };
-
-
   return (
-    <UserContext.Provider value={{ userData, getUserFromId, favourites, getPublicUserFromId }}>
+    <UserContext.Provider value={{ userData, getUserFromId, favourites }}>
       {children}
     </UserContext.Provider>
   );
