+++ src/components/pubprofilecomp/PubProfileComp.jsx
@@ -1,15 +1,14 @@
 import React from "react";
 import "./PubProfileComp.css";
-import { UserContext } from "../context/UserContext";
+import { PubUserContext } from "../context/PubUserContext";
 import { useContext, useState, useEffect } from "react";
 import { Link, useParams } from "react-router-dom";
 import { SearchContext } from "../context/SearchContext";
 import { AuctionContext } from "../context/AuctionContext";
-import { ReviewContext } from "../context/ReviewContext";
 
 export const PubProfileComp = () => {
   const inputRef = React.useRef(null);
-  const { userData, getPublicUserFromId } = useContext(UserContext);
+  const { userData, getPublicUserFromId } = useContext(PubUserContext);
   const { userId } = useParams(SearchContext);
   const { usersAuctions, fetchUsersAuctions } = useContext(AuctionContext);
